feat: Expose event id functions
