+++ apps/hubble/src/console/console.ts
@@ -9,6 +9,7 @@ import {
   fromFarcasterTime,
   bytesToHexString,
   hexStringToBytes,
+  makeEventId,
 } from "@farcaster/hub-nodejs";
 import path from "path";
 import * as repl from "repl";
@@ -21,6 +22,7 @@ import { RpcClientCommand } from "./rpcClientCommand.js";
 import { WarpcastTestCommand } from "./warpcastTestCommand.js";
 import { SyncId } from "../network/sync/syncId.js";
 import { TrackHubDelayCommand } from "./trackHubDelayCommand.js";
+import { extractEventTimestamp } from "@farcaster/core";
 
 export const DEFAULT_RPC_CONSOLE = "127.0.0.1:2283";
 
@@ -98,6 +100,8 @@ export const startConsole = async (addressString: string, useInsecure: boolean) 
   replServer.context["fromFarcasterTime"] = fromFarcasterTime;
   replServer.context["bytesToHex"] = bytesToHexString;
   replServer.context["hexToBytes"] = hexStringToBytes;
+  replServer.context["extractEventTimestamp"] = extractEventTimestamp;
+  replServer.context["makeEventId"] = makeEventId;
 
   // Run the info command to start
